container-timelapse/requirements.txt
fastapi
uvicorn
pillow
numpy
pandas
matplotlib
seaborn
contextily
cartopy
pyproj
pymongo
pytz
tzlocal